requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-multipart
PyJWT
passlib
bcrypt
SQLAlchemy